Fix option thing

+++ src/project/foodorder/actions/AddOptionAction.java
@@ -1,13 +1,14 @@
 package project.foodorder.actions;
-import project.core.keyboard.Keyboard;
+
 import java.util.Date;
+
+import project.core.keyboard.Keyboard;
 import project.core.menu.MenuItem;
 import project.foodorder.AplicationSession;
 import project.foodorder.model.AvailableOption;
 
 public class AddOptionAction extends MenuItem {
 	private Keyboard keyboard = Keyboard.getInstance();
-
 
 	public AddOptionAction() {
 		super("1", "Add");
@@ -17,10 +18,9 @@ public class AddOptionAction extends MenuItem {
 	public void doAction() {
 		String id = keyboard.getString("ID: ");
 		Date data = keyboard.getDate("Data: ");
-		String details = keyboard.getString("Details: ");	
-		double price = keyboard.getInt("Price: ");
-		
-		
+		String details = keyboard.getString("Details: ");
+		double price = keyboard.getDouble("Price: ");
+
 		AvailableOption option = new AvailableOption(id, price, details, data);
 		AplicationSession.getInstance().getDataBase().addOption(option);
 	}
